+++ src/test/java/io/github/vitalijr2/ytimebot/youtube/YouTubeTimeServiceSlowTest.java
@@ -8,6 +8,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import feign.mock.HttpMethod;
 import feign.mock.MockClient;
+import io.github.vitalijr2.ytimebot.youtube.VideoData.PrivacyStatus;
+import io.github.vitalijr2.ytimebot.youtube.VideoData.Thumbnail;
+import io.github.vitalijr2.ytimebot.youtube.VideoData.UploadStatus;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -31,18 +34,14 @@ class YouTubeTimeServiceSlowTest {
   }
 
   @DisplayName("Happy path")
-  @ParameterizedTest(name = "{6} - {2}")
-  @CsvSource({"src/test/resources/localized_happy_path.json,Channel title,Localized description,"
-      + "https://i.ytimg.com/vi/qwerty/sddefault.jpg,public,"
-      + "https://i.ytimg.com/vi/qwerty/default.jpg,Localized title,processed",
-      "src/test/resources/original_happy_path.json,Channel title,Original description,"
-          + "https://i.ytimg.com/vi/qwerty/sddefault.jpg,public,"
-          + "https://i.ytimg.com/vi/qwerty/default.jpg,Original title,processed"})
-  void happyPath(String resourcePath, String channelTitle, String description, String preview,
-      String privacyStatus, String thumbnail, String title, String uploadStatus)
-      throws IOException {
+  @ParameterizedTest(name = "{2} - {1}")
+  @CsvSource({"src/test/resources/localized_happy_path.json,Localized description,Localized title",
+      "src/test/resources/original_happy_path.json,Original description,Original title"})
+  void happyPath(String resourcePath, String description, String title) throws IOException {
     // given
     var responseBody = Files.readString(Path.of(resourcePath), ISO_8859_1);
+    var preview = new Thumbnail("https://i.ytimg.com/vi/qwerty/sddefault.jpg", 480, 640);
+    var thumbnail = new Thumbnail("https://i.ytimg.com/vi/qwerty/default.jpg", 90, 120);
 
     mockClient.add(HttpMethod.GET, "http://api.youtube.test/youtube/v3/videos?"
         + "hl=en&part=snippet&part=status&id=qwerty&key=qwerty-xyz", 200, responseBody);
@@ -52,22 +51,23 @@ class YouTubeTimeServiceSlowTest {
 
     // then
     assertAll("Localized data", () -> assertNull(videoData.errorReason(), "error reason"),
-        () -> assertEquals(channelTitle, videoData.channelTitle(), "channel title"),
+        () -> assertEquals("Channel title", videoData.channelTitle(), "channel title"),
         () -> assertEquals(description, videoData.description(), "description"),
         () -> assertEquals(preview, videoData.preview(), "preview"),
-        () -> assertEquals(privacyStatus, videoData.privacyStatus(), "privacy status"),
+        () -> assertEquals(PrivacyStatus.Public, videoData.privacyStatus(), "privacy status"),
         () -> assertEquals(thumbnail, videoData.thumbnail(), "thumbnail"),
         () -> assertEquals(title, videoData.title(), "title"),
-        () -> assertEquals(uploadStatus, videoData.uploadStatus(), "upload status"));
+        () -> assertEquals(UploadStatus.Processed, videoData.uploadStatus(), "upload status"));
   }
 
   @DisplayName("Without host language")
   @Test
-  void withoutHostLanguage()
-      throws IOException {
+  void withoutHostLanguage() throws IOException {
     // given
     var responseBody = Files.readString(Path.of("src/test/resources/localized_happy_path.json"),
         ISO_8859_1);
+    var preview = new Thumbnail("https://i.ytimg.com/vi/qwerty/sddefault.jpg", 480, 640);
+    var thumbnail = new Thumbnail("https://i.ytimg.com/vi/qwerty/default.jpg", 90, 120);
 
     mockClient.add(HttpMethod.GET, "http://api.youtube.test/youtube/v3/videos?"
         + "part=snippet&part=status&id=qwerty&key=qwerty-xyz", 200, responseBody);
@@ -79,13 +79,32 @@ class YouTubeTimeServiceSlowTest {
     assertAll("Localized data", () -> assertNull(videoData.errorReason(), "error reason"),
         () -> assertEquals("Channel title", videoData.channelTitle(), "channel title"),
         () -> assertEquals("Localized description", videoData.description(), "description"),
-        () -> assertEquals("https://i.ytimg.com/vi/qwerty/sddefault.jpg", videoData.preview(),
-            "preview"),
-        () -> assertEquals("public", videoData.privacyStatus(), "privacy status"),
-        () -> assertEquals("https://i.ytimg.com/vi/qwerty/default.jpg", videoData.thumbnail(),
-            "thumbnail"),
+        () -> assertEquals(preview, videoData.preview(), "preview"),
+        () -> assertEquals(PrivacyStatus.Public, videoData.privacyStatus(), "privacy status"),
+        () -> assertEquals(thumbnail, videoData.thumbnail(), "thumbnail"),
         () -> assertEquals("Localized title", videoData.title(), "title"),
-        () -> assertEquals("processed", videoData.uploadStatus(), "upload status"));
+        () -> assertEquals(UploadStatus.Processed, videoData.uploadStatus(), "upload status"));
+  }
+
+  @DisplayName("Thumbnails without size")
+  @Test
+  void thumbnailsWithoutSize() throws IOException {
+    // given
+    var responseBody = Files.readString(Path.of("src/test/resources/thumbnails_without_size.json"),
+        ISO_8859_1);
+    var preview = new Thumbnail("https://i.ytimg.com/vi/qwerty/sddefault.jpg", null, null);
+    var thumbnail = new Thumbnail("https://i.ytimg.com/vi/qwerty/default.jpg", null, null);
+
+    mockClient.add(HttpMethod.GET, "http://api.youtube.test/youtube/v3/videos?"
+        + "part=snippet&part=status&id=qwerty&key=qwerty-xyz", 200, responseBody);
+
+    // when
+    var videoData = service.getLinkData("https://youtu.be/qwerty", null);
+
+    // then
+    assertAll("Localized data", () -> assertNull(videoData.errorReason(), "error reason"),
+        () -> assertEquals(preview, videoData.preview(), "preview"),
+        () -> assertEquals(thumbnail, videoData.thumbnail(), "thumbnail"));
   }
 
   @DisplayName("Error")